+++ Source/missiles.h
@@ -17,7 +17,7 @@ extern MissileStruct missile[MAXMISSILES];
 extern int nummissiles;
 
 void GetDamageAmt(int sn, int sl, int* mind, int* maxd);
-unsigned CalcMonsterDam(uint16_t mor, BYTE mRes, unsigned mindam, unsigned maxdam, bool penetrates);
+unsigned CalcMonsterDam(unsigned mor, BYTE mRes, unsigned mindam, unsigned maxdam, bool penetrates);
 unsigned CalcPlrDam(int pnum, BYTE mRes, unsigned mindam, unsigned maxdam);
 int CheckMonCol(int _mnum_);
 int CheckPlrCol(int _pnum_);
